Fix initialization bug that cause app not to check for updates, add queue to logging

+++ OpenIPC_Config/App.axaml.cs
@@ -28,14 +28,12 @@ public class App : Application
     public static IServiceProvider ServiceProvider { get; private set; }
 
     public static string OSType { get; private set; }
-    
-    
 
 #if DEBUG
     private bool _ShouldCheckForUpdates = false;
 #else
     private bool _ShouldCheckForUpdates = true;
-#endif    
+#endif
 
     private void DetectOsType()
     {
@@ -60,13 +58,13 @@ public class App : Application
         var configPath = GetConfigPath();
 
         // Create default settings if not present
-        //if (!File.Exists(configPath))
-        //{
+        if (!File.Exists(configPath))
+        {
             // create the file
             var defaultSettings = createDefaultAppSettings();
             File.WriteAllText(configPath, defaultSettings.ToString());
-            Log.Information($"Default appsettings.json created at {configPath}");
-        //}
+            Log.Information($"Default appsettings.json created at {configPath}"); // Log here, using a basic config
+        }
 
         Console.WriteLine($"Loading configuration from: {configPath}");
         // Build configuration
@@ -78,42 +76,69 @@ public class App : Application
         return configuration;
     }
 
-    private void ReconfigureLogger(IConfiguration configuration)
+    private void ConfigureServices(IServiceCollection services, IConfiguration configuration)
     {
-        var eventAggregator = ServiceProvider.GetRequiredService<IEventAggregator>();
+        // Register IEventAggregator as a singleton
+        services.AddSingleton<IEventAggregator, EventAggregator>();
+        services.AddSingleton<IEventSubscriptionService, EventSubscriptionService>();
+        services.AddSingleton<ISshClientService, SshClientService>();
+        services.AddSingleton<IMessageBoxService, MessageBoxService>();
 
-        //Log.Logger = null;
-        Log.Logger = new LoggerConfiguration()
-            .ReadFrom.Configuration(configuration)
-            //.WriteTo.Console() // Keep console logging
-            .WriteTo.Sink(new EventAggregatorSink(eventAggregator)) // Add EventAggregatorSink
-            .CreateLogger();
+        services.AddSingleton<IYamlConfigService, YamlConfigService>();
+        
+        // Register IConfiguration
+        services.AddSingleton<IConfiguration>(configuration);
+        services.AddTransient<DeviceConfigValidator>();
+        services.AddSingleton<IGlobalSettingsService, GlobalSettingsService>();
+
+        // Register IConfiguration
+        services.AddTransient<DeviceConfigValidator>();
+
+        services.AddSingleton<HttpClient>();
+        services.AddSingleton<IGitHubService, GitHubService>();
+
+        // add memory cache
+        services.AddMemoryCache();
+
+        // Register ViewModels
+        RegisterViewModels(services);
+
+        // Register Views
+        RegisterViews(services);
+
+        // Register Logger using factory
+        services.AddSingleton<ILogger>(sp =>
+        {
+            var eventAggregator = sp.GetRequiredService<IEventAggregator>();
+            return new LoggerConfiguration()
+                .ReadFrom.Configuration(configuration)
+                //.WriteTo.Console() // Keep console logging
+                .WriteTo.Sink(new EventAggregatorSink(eventAggregator)) // Add EventAggregatorSink
+                .CreateLogger();
+        });
+    }
+
+    public override void OnFrameworkInitializationCompleted()
+    {
+        // Step 1: Load configuration
+        var configuration = LoadConfiguration();
+
+        // Step 2: Configure DI container
+        var serviceCollection = new ServiceCollection();
+        ConfigureServices(serviceCollection, configuration);
+        ServiceProvider = serviceCollection.BuildServiceProvider();
+
+        // Step 3: Initialize logger (resolve it from service provider)
+        Log.Logger = ServiceProvider.GetRequiredService<ILogger>();
 
         Log.Information(
             "**********************************************************************************************");
         Log.Information($"Starting up log for OpenIPC Configurator v{VersionHelper.GetAppVersion()}");
         Log.Information("Logger initialized successfully.");
-    }
-
-    public override void OnFrameworkInitializationCompleted()
-    {
-        // Step 1: Initialize basic logger 
-        // not sure if this is needed
-        //InitializeBasicLogger();
-
-        // Step 2: Load configuration
-        var configuration = LoadConfiguration();
-
-        // Configure DI container
-        var serviceCollection = new ServiceCollection();
-        ConfigureServices(serviceCollection, configuration);
-        ServiceProvider = serviceCollection.BuildServiceProvider();
-
-        // Step 4: Reconfigure logger with DI services
-        ReconfigureLogger(configuration);
+        Log.Information("Starting up....");
 
         // check for updates
-        if(_ShouldCheckForUpdates)
+        if (_ShouldCheckForUpdates)
             CheckForUpdatesAsync();
 
         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
@@ -167,7 +192,6 @@ public class App : Application
 
         return configPath;
     }
-    
 
     public virtual async Task ShowUpdateDialogAsync(string releaseNotes, string downloadUrl, string newVersion)
     {
@@ -200,7 +224,7 @@ public class App : Application
             Assembly.GetExecutingAssembly().GetName().Name, "appsettings.json");
 
         Console.WriteLine($"Loading configuration from: {configPath}");
-        
+
         // Create an IConfiguration instance
         var configuration = new ConfigurationBuilder()
             .AddJsonFile(configPath, false, true)
@@ -241,43 +265,10 @@ public class App : Application
         }
     }
 
-    private void ConfigureServices(IServiceCollection services, IConfiguration configuration)
-    {
-        // Register IEventAggregator as a singleton
-        services.AddSingleton<IEventAggregator, EventAggregator>();
-        services.AddSingleton<IEventSubscriptionService, EventSubscriptionService>();
-        services.AddSingleton<ISshClientService, SshClientService>();
-        services.AddSingleton<IMessageBoxService, MessageBoxService>();
-
-        services.AddSingleton<IYamlConfigService, YamlConfigService>();
-        services.AddSingleton<ILogger>(sp => Log.Logger);
-        
-        // Register IConfiguration
-        services.AddSingleton<IConfiguration>(configuration);
-        services.AddTransient<DeviceConfigValidator>();
-        services.AddSingleton<IGlobalSettingsService, GlobalSettingsService>();
-        
-        // Register IConfiguration
-        services.AddTransient<DeviceConfigValidator>();
-
-        services.AddSingleton<HttpClient>();
-        services.AddSingleton<IGitHubService, GitHubService>();
-
-        // add memory cache
-        services.AddMemoryCache();
-
-        // Register ViewModels
-        RegisterViewModels(services);
-
-        // Register Views
-        RegisterViews(services);
-    }
-
     private static void RegisterViewModels(IServiceCollection services)
     {
         // Register ViewModels
         services.AddSingleton<MainViewModel>();
-        
 
         services.AddSingleton<CameraSettingsTabViewModel>();
         services.AddSingleton<ConnectControlsViewModel>();
@@ -290,7 +281,6 @@ public class App : Application
         services.AddSingleton<WfbTabViewModel>();
         services.AddSingleton<FirmwareTabViewModel>();
         services.AddSingleton<PresetsTabViewModel>();
-        
     }
 
     private static void RegisterViews(IServiceCollection services)
@@ -309,12 +299,10 @@ public class App : Application
         services.AddTransient<FirmwareTabView>();
         services.AddTransient<WfbTabView>();
         services.AddTransient<PresetsTabView>();
-        
     }
 
     private JObject createDefaultAppSettings()
     {
-
         string logPath = Path.Combine(OpenIPC.AppDataConfigDirectory, "Logs", "configurator.log");
 
         // Create default settings
